+++ CountlyConnectionManager.m
@@ -606,9 +606,9 @@ static dispatch_once_t onceToken;
     }
 #endif
 
-    if([Countly.user hasUnsyncedChanges])
+    if ([CountlyUserDetails.sharedInstance hasUnsyncedChanges])
     {
-        [Countly.user save];
+        [CountlyUserDetails.sharedInstance save];
     }
 
     isSessionStarted = YES;
@@ -649,9 +649,9 @@ static dispatch_once_t onceToken;
         return;
     }
     
-    if([Countly.user hasUnsyncedChanges])
+    if ([CountlyUserDetails.sharedInstance hasUnsyncedChanges])
     {
-        [Countly.user save];
+        [CountlyUserDetails.sharedInstance save];
     }
 
     NSString* queryString = [[self queryEssentials] stringByAppendingFormat:@"&%@=%d",
@@ -691,9 +691,9 @@ static dispatch_once_t onceToken;
 
 - (void)sendEventsWithSaveIfNeeded
 {
-   if([Countly.user hasUnsyncedChanges])
+    if ([CountlyUserDetails.sharedInstance hasUnsyncedChanges])
     {
-        [Countly.user save];
+        [CountlyUserDetails.sharedInstance save];
     }
     else
     {